fix reference issues and move the generic class EventMesh to the Common module
