+++ eventmesh-examples/src/main/java/org/apache/eventmesh/grpc/pub/eventmeshmessage/RequestReplyInstance.java
@@ -18,7 +18,7 @@
 package org.apache.eventmesh.grpc.pub.eventmeshmessage;
 
 import org.apache.eventmesh.client.grpc.producer.EventMeshGrpcProducer;
-import org.apache.eventmesh.client.tcp.common.EventMeshCommon;
+import org.apache.eventmesh.common.EventMeshCommon;
 import org.apache.eventmesh.common.ExampleConstants;
 import org.apache.eventmesh.common.utils.ThreadUtils;
 import org.apache.eventmesh.grpc.GrpcAbstractDemo;
@@ -45,7 +45,7 @@ public class RequestReplyInstance extends GrpcAbstractDemo {
 
             for (int i = 0; i < MESSAGE_SIZE; i++) {
                 eventMeshGrpcProducer.requestReply(buildEventMeshMessage(content,
-                    ExampleConstants.EVENTMESH_GRPC_RR_TEST_TOPIC),
+                        ExampleConstants.EVENTMESH_GRPC_RR_TEST_TOPIC),
                     EventMeshCommon.DEFAULT_TIME_OUT_MILLS);
                 ThreadUtils.sleep(1, TimeUnit.SECONDS);
             }
